chore: splitting logic
