+++ src/text/characters.ts
